refactor: 알림 관련 API 및 로직 리팩토링 #234
