+++ src/main/java/com/ceos/bankids/service/UserService.java
@@ -1,12 +1,13 @@
 package com.ceos.bankids.service;
 
-import com.ceos.bankids.mapper.request.ExpoRequest;
-import com.ceos.bankids.mapper.request.UserTypeRequest;
 import com.ceos.bankids.domain.User;
 import com.ceos.bankids.dto.LoginDTO;
 import com.ceos.bankids.dto.MyPageDTO;
 import com.ceos.bankids.dto.OptInDTO;
 import com.ceos.bankids.dto.UserDTO;
+import com.ceos.bankids.mapper.request.ExpoRequest;
+import com.ceos.bankids.mapper.request.UserTypeRequest;
+import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 
@@ -36,4 +37,7 @@ public interface UserService {
     public OptInDTO updateServiceOptIn(User user);
 
     public OptInDTO getOptIn(User user);
+
+    public List<User> readAllUserList();
+
 }
